requirements.txt
asgiref==3.6.0
crispy-bootstrap4==2022.1
crispy-bootstrap5==0.7
Django==4.1.9
django-crispy-forms==2.0
django-extensions==3.2.1
django-filter==23.2
djangorestframework==3.14.0
djangorestframework-jsonapi==6.0.0
inflection==0.5.1
python-decouple==3.8
python-dotenv==1.0.0
pytz==2023.3
sqlparse==0.4.4
tzdata==2023.3
